refactor(silencer): split SDL/input/gamepad code into events.cpp
Phase 4 mechanical move. HandleSDLEvents, OnScancodeDown/Up,
UpdateInputState, OpenFirstGamepad, PollGamepadState, and the
INPUT_FIELDS table now live in src/game/events.cpp. Declarations
stay on Game in game.h. No behavior change.
